backend/pyproject.toml
[tool.poetry]
name = "ai-c-suite-backend"
version = "0.1.0"
description = "FastAPI Backend for AI C-Suite Agent SaaS"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
pinecone-client = "^3.0.0"
openai = "^1.12.0"
crewai = "^0.1.0" # Check for latest version
httpx = "^0.26.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
